fixed prio issue

+++ stm32f4xx_drivers3/drivers/Inc/stm32f407xx_gpio.h
@@ -87,6 +87,6 @@ void GPIO_WriteToOutputPort(GPIO_RegDef_t *pGPIOx, uint16_t val);
 void GPIO_ToggleOutputPin(GPIO_RegDef_t *pGPIOx, uint8_t PinNum);
 void GPIO_IRQITConfig(uint8_t IRQNum, uint8_t EnorDi);
 void GPIO_IRQHandling(uint8_t PinNum);
-void GPIO_IRQPriorityConfig(uint8_t IRQNum, uint8_t IRQPriority);
+void GPIO_IRQPriorityConfig(uint8_t IRQNum, uint32_t IRQPriority);
 
 #endif /* INC_STM32F407XX_GPIO_H_ */
